requirements.txt
ttkbootstrap
beautifulsoup4
requests
pillow
python-docx
PyQt6
qtawesome
